Sync Watermark: Use deterministic techniques.
Previously, we were using deadlines based to notify synced watermarks that an
index is pending. The problem this caused was that on slower machines,
the deadline would happen before the index had a chance to be applied
and figure if it had something to be synced. This would cause the
watermark to advance, and then an older index to be received, causing an
assert failure.

Now, we treat sync just like we do applied, keep tracking of all
possible exit paths and handling them accordingly.

This change also includes renaming variables and functions to make the
terminology clear.