Fix bugs in bulk loader. Add Dgraph debug tool. (#2449)
When building multiple shards, we want each predicate to only lie to one shard, and nowhere else. Bulk loader was doing that, but then adding the schema for all predicates on all shards, which causes confusion about ownership of the predicates. This changes that to only add schema for the predicates that the shard holds.

Similarly, when outputting _predicate_ edge, the shard being used was the one corresponding to the original predicate. Instead, we should use the shard corresponding to _predicate_.

This PR fixes #2129 .

Added a new debug tool which can iterate over posting store and spit out stats per predicate. Useful for debugging.