Fix up predicate snapshot (#2487)
Fix up both the way we create a snapshot and the way we stream it. Make the logic robust and add tests to ensure that the calculation of snapshot is right.

Streaming no longer needs to worry about `min_ts`. We no longer send a snapshot upfront. Raft would request it as it desires, and the follower sends the snapshot details it needs from the leader. The leader in turn checks if its snapshot matches with what the follower wants. This is a better logic, but not entirely foolproof (comment in predicate.go explains).

Bring in Badger changes, which add a `DropAll` method, which is called before applying the snapshot.