Decrease the number of aborts by using SPO for entity to entity connections. (#2556)
This change would make the conflict detection lot more precise. For SPO connections, where O = UID, it would use all 3 for conflict detection. For SPV, where V = value, it would use SP for conflict detection.

This significantly reduces the number of aborts that happen in the system and allows a lot more concurrency.