Update dgraph-ha.yaml to run one group instead of two.
This HA configuration gets a lot of use in the community. But it
immediately starts users off with a sharded cluster when the data set
probably doesn't warrant sharding.

This change reduces the number of Dgraph Servers from 6 to 3. This is
still an HA setup with 3 Zeros and 3 Servers. For comparison, the
dgraph-multi.yaml config sets up 1 Zero and 3 Servers.

This change also adds comment headers to dgraph-multi.yaml and
dgraph-ha.yaml describing the availability characteristics per config.