Refactor index rebuild process (#2650)
The index rebuilding code was a mess, copied many times over for each use case. Refactor all that into one "framework", which takes in function holding the core logic, while removing the boilerplate out of each invocation (similar to how orchestrate works).

The index building is done via a single txn object, which now holds all the mutations in memory. Once the rebuild is done, the posting lists are rolled up and flushed to disk at the timestamp provided. Honestly, this simplification took a while to occur to me. To avoid polluting the LRU cache, txn struct now exposes a getList API, which defaults to posting.Get, but the rebuild code above uses a local cache; so we can isolate the rebuild proces.

Also, posting list iterate which takes in a function, that function should really be returning an error, so it can be sent back to the caller; instead of the caller doing their own error catching outside of the function call.