Switch Dgraph back to Apache 2.0 (#2652)
While the promise of Commons Clause is a great one, i.e. restricting resale while allowing proprietary usage, its subjective implementation generates a lot of friction and confusion for the open source community.

Being an open source developer at heart, I'm glad to be bringing Dgraph back to Apache 2.0 license. It is something that we started off with, but then got lost into the AGPL woods. At the end of the day, Apache license is the gold standard of licenses, and something open source users (i.e. the world) feel very comfortable with. The vision for Dgraph is to bring graph databases mainstream, allowing them to be the primary databases holding the source of truth, and that's only possible if it can be adopted by any company, big or small, without any concerns.

**Apache license lets achieve that vision.**

We will be adding proprietary features later on under an enterprise license, but as long as I'm in charge, we're not moving the core away from Apache 2.0 license. This is it.