Replace bp128 with a simple to understand codec (#2719)
We had been using a SIMD based integer encoder, which was pulled in from an abandoned repository, with no maintenance and no testing. The code was complex and no one touched it since it was put in place.

This PR replaces that encoding scheme using the same concepts, via protocol buffers and Varint encoding - both of which are solid and reliable. The new scheme of encoding is a lot slower than the previous one, but is simple to understand and test. It does not show weird panic issues we've been seeing. So, some point later, we'd want to come back to this and try to speed things up.

Fixes #2470 .

* Add Encoder and Decoder to make working with UidPack easier. Remove UidPostings bit.
* Replaced some instances of bp128 with the new encoding. Need to work on algo package as well.
* Switched storage of deltas to a byte slice.
* Rename binary to codec
* Break everything by removing PostingList.Uids
* Update algo pkg with the new codec.
* Update debug tool.
* Switch count indexer to use x.TxnWriter.
* Add benchmarks (things are slow. sigh. But, simplicity and reliability over performance for now.)