Optimize has function (#2724)
Since the last change to make `has(...)` function to work correctly with transactions, things have been slow due to the need to build the entire posting list to determine if it has a posting or not. This PR optimizes that by utilizing BitCompletePosting. If this bit is found, then there must be a valid posting in the list, so we don't need to recreate the entire list.

This speeds up `handleHasFunction` considerably, for all predicates which don't have `@lang` index. For the predicates which have the lang index set, they're slowed down by `filterStringFunction`, but there's no easy fix there. Added a TODO for that.

Partially fixes #2533 for preds which don't have `@lang` index.

Changelog:

* No need to retrieve posting list if we have the complete bit set. It must have at least some elements to be present.
* Understand the cause of slowdown, i.e. filterStringFunction. Improve IsEmpty func in posting/list.go