Report unhealthy quickly
If we see a Raft node having TransientFailure, we should report it
quickly and mark the connection as unhealthy. This avoids us trying to
send a lot of messages to that server. We also report that node as
Unhealthy, to indicated this fact to Raft. I tested with docker,
stopping and restarting a node. After a restart, the node reconnects
successfully with the cluster and receives proposals.

Fixes #2702.