Set the Prefix option during iteration (#2780)
Badger now has a new way to iterate over a certain prefix. This is done by setting a new `Prefix` option in `IteratorOptions`. This option makes it faster to Seek over the iterator, by not including any table which does not overlap with the provided prefix.

Also, Badger has a `NewKeyIterator`, which additionally uses the bloom filters to further remove tables from the candidate set, during iterator creation. Use that during `newList`.

Commits:

* Bring in Badger's prefix change during iterator creation.
* Use NewKeyIterator during posting list creation.