Automatically pick up the k8s pod namespace for DNS resolution.
The k8s DNS names contain the name of the namespace in which the pods
are deployed. The configs assumed that Dgraph is running in the
default namespace. This doesn't work when the config is deployed to
the non-default namespace.

By injecting the namespace name into the pod, we can automatically
use the correct namespace for DNS names in the command.