Fix Rollup to pick max update commit ts
If the STAR delete marker is the most recent update in a posting list,
Rollup ignores the version corresponding to the delete marker. This PR
fixes that by iterating over the postings and the list to pick the
highest commit timestamp. Also add tests to test this and to test the
handling of BitEmptyPosting and BitCompletePosting.