Snapshot streaming:
Typically a follower would ask a leader to send a snapshot. There's a
rare edge that we witnessed where the leader would switch after asking
the follower to get a snapshot. In that case, if the follower sends a
snapshot request to another follower, Dgraph ends up in an infinitely
failing loop.

We skip this loop by not checking if the receiver of the request is a
leader. Any follower should be able to service the request, once it is
past the read timestamp. That's what this PR does.