Support comments in schema (#3133)
* schema/state.go: add comment state to lexer

This adds comment detection to lexer. A comment starts with # and ends with
new line or EOF. When new line is detected, it's emitted so the caller can handle it.
If EOF is found, that should end further parsing.

* schema/parse_test.go: add test for comments in schema types

* gql/parser.go: fixed typo

* schema/state.go: rename func lexComment to lexTextComment

Package gql has another function called lexComment that works slightly
different to this one, so I renamed this one to avoid confusion.