Do not retrieve facets when max recurse depth has been reached. (#3190)
Currently, recurse will retrive the facets at max depth, which results
in the facets being present in the results while the node on the other
side of the edge is absent from the result. This change modifies the
logic so that facets are only included if the node to which the edge
connects is included in the results as well.