Use UTC Hour, Day, Month, Year for comparison
The time.Hour() function, for example, returns the value
of hour stored, without accounting for timezone information.
E.g. for timestamp, 2019-05-28T14:41:57+30:00, the Hour value
would be 14 whereas when we use this value for comparison,
that would not correctly compare two timestamps.

Now, we use UTC value of hour for creating index and for
comparison to ensure that we correctly compare all timestamps
having timezone information.

Note that all the timestmaps are parsed assuming UTC timezone
if no timezone is specified. This is the default behaviour
in Go for the time.Parse function as well.