During startup, don't upsert initial schema if it already exists. (#3374)
* During startup, don't upsert initial schema if it already exists.

Currently, there are lots of schema insertions during startup. This
change optimizes the insertion of the schema for the initial/reserved
predicates by only doing it if the predicate is not currently being
served by any group or the definition in storage is different from the
proposed schema.

Tested by starting and stoping a container and looking at the logs.
Without this change, a lot of logs about schema changes are printed
after the stopped container is restarted. With the change, the amount of
logs is greatly reduced.