Enterprise - Support applying a license using /enterpriseLicense endpoint in Zero (#3824)
This change gets rid of the `--enterprise_features` flag in Alpha. 

All users of the non-oss binary would get enterprise features enabled by default for 30 days.  After the 30 days, users who want to use the enterprise features, would have to obtain a license from us and supply it to a Zero node using the `/enterpriseLicense` endpoint. The license is a message that's signed using our private key and is verified using the public key in the code.The license information would automatically be propagated to Alpha nodes along with other membership info. After the expiry of the license, the cluster keeps on working but ACL and Backup would stop working until the license is renewed.