Fix bug in bulk loader when store_xids is true(#3950)
In mapper stage of bulk loader, NQuads are processed by calling processNQuad(). If store_xids is true, we create extra NQuad for xid and uid mapping and call processNQuad(). Ideally this should be done only once. But we were not tracking if we had already created NQuad or not. This was resulting in multiple recursive calls to processNQuad() and leading to overflowing of stack.
This PR fixes it by returning true/false from AssignUid() of xidmap. When it returns false, we know its an old Uid and avoid creating extra NQuad and recursive call to processNQuad(). 

Fixes: #3922