Bulk loader allocates reserved predicates in first reduce shard. (#4202)
This PR contains a couple of related changes.

Bulk loader forces reserved predicates to end up in the first reduce shard. It
also writes a file in the posting directories with the proposed group ID for
each shard.
Dgraph looks at that file during startup and uses it to request the right group
ID from zero.
The change is being tested by modifying the 21million test to use multiple groups
and add a new query to verify the number of nodes with a dgraph.type predicate.
If the test runs without the fix, dgraph.type sometime ends up in a different group.

Fixes #3968.