Use txn writer to write schema postings (#4296)
Fixes #3916

Although I was not able to reproduce this. Stacktrace suggests, while writing schema to Badger, it 
got crashed with error: Txn is too big to fit into one request. This might be because, we are using single transaction to write schema for all predicates.
This PR, replaces single transaction with TxnWriter to write schema.