Make uid function work with value variables in upsert blocks (#4425)
* Make uid function work with value variables in upsert blocks

Currently, only variables containing UIDs (e.g. `uid as uid`) work correctly
when used in upsert blocks with the `uid` function.

With this fix, all variables (e.g. `value as value`) work appropriately, as we
fetch the list of UIDs from the keys in the value map when needed.

Fixes #4424

* added tests for upsert with value variables