GraphQL: Support reverse predicates as arguments to dgraph directive. (#4676)
This change allows supplying a reverse field as a dgraph predicate. For e.g.

```
type Movie {
        id: ID!
        name: String!
        director: [Director] @dgraph(pred: "~directed.movies")
}

type Director {
        id: ID!
        name: String!
        directed: [Movie] @dgraph(pred: "directed.movies")
}
```

Mutations are not allowed along the reverse direction and are only allowed in the forward direction. Queries are allowed along both the directions. We also don't allow specifying @hasInverse along with reverse edge on a field as its not required anymore because Dgraph would handles inverses.