Support ACL operations using the admin GraphQL API (#4760)
This PR supports all the operations currently supported by the ACL tool using the admin API available at /admin. The current tests in acl_test.go have also been modified to use the new API. The dgraph acl tool would be removed as part of another PR.

Some more testing is required for the API to test the behavior when users who are not part of the guardian group use it. I'll do that as part of a separate change.