graphql: Fix deleteGroup and deleteUser mutations. (#4820)
deleteGroup should also delete the group from the user group lists that it belongs to. This was not happening before because we were not adding the delete mutation like we do for fields which have the @hasInverse directive. This happens now.

deleteUser should also delete the user from the groups that it belongs to. This was not happening because Dgraph didn't have the type info for User type and wasn't correctly deleting its edges. Now the type information is added to the list of initial types.